+++ src/conf.c
@@ -134,6 +134,24 @@ static void conf_srv_location(struct srv *srv, cfg_t *cfg)
 	}
 }
 
+static void conf_srv_proxy(struct srv *srv, cfg_t *cfg)
+{
+	size_t i;
+
+	for (i = 0; i < cfg_size(cfg, "proxy-pass") && i < NELEMS(srv->proxy); i++) {
+		cfg_t *proxy;
+
+		proxy = cfg_getnsec(cfg, "proxy-pass", i);
+		if (!proxy)
+			return;
+
+		srv->proxy[i].pattern  = (char *)cfg_title(proxy);
+		srv->proxy[i].vhost    = cfg_getstr(proxy, "host");
+		srv->proxy[i].backend  = cfg_getstr(proxy, "backend");
+		srv->proxy[i].redirect = cfg_getstr(proxy, "proxy-redirect");
+	}
+}
+
 static void conf_ssl(struct srv *srv, cfg_t *cfg)
 {
 	cfg_t *ssl;
@@ -198,6 +216,7 @@ int conf_srv(struct srv arr[], size_t len)
 		conf_ssl(&arr[i], srv);
 		conf_redirect(&arr[i], srv);
 		conf_srv_location(&arr[i], srv);
+		conf_srv_proxy(&arr[i], srv);
 	}
 
 	return (int)i;
@@ -207,6 +226,12 @@ static int read_config(char *fn)
 {
 	cfg_opt_t location_opts[] = {
 		CFG_STR ("path", NULL, CFGF_NONE),
+		CFG_END ()
+	};
+	cfg_opt_t proxy_opts[] = {
+		CFG_STR ("host",           NULL, CFGF_NONE),
+		CFG_STR ("backend",        NULL, CFGF_NONE),
+		CFG_STR ("proxy-redirect", NULL, CFGF_NONE),
 		CFG_END ()
 	};
 	cfg_opt_t redirect_opts[] = {
@@ -241,12 +266,13 @@ static int read_config(char *fn)
 		CFG_END ()
 	};
 	cfg_opt_t server_opts[] = {
-		CFG_STR ("hostname", hostname, CFGF_NONE),
-		CFG_INT ("port",     port, CFGF_NONE),
-		CFG_STR ("path",     path, CFGF_NONE),
-		CFG_SEC ("location", location_opts, CFGF_MULTI | CFGF_TITLE),
-		CFG_SEC ("ssl",      ssl_opts, CFGF_MULTI),
-		CFG_SEC ("redirect", redirect_opts, CFGF_MULTI | CFGF_TITLE),
+		CFG_STR ("hostname",   hostname, CFGF_NONE),
+		CFG_INT ("port",       port, CFGF_NONE),
+		CFG_STR ("path",       path, CFGF_NONE),
+		CFG_SEC ("location",   location_opts, CFGF_MULTI | CFGF_TITLE),
+		CFG_SEC ("ssl",        ssl_opts, CFGF_MULTI),
+		CFG_SEC ("redirect",   redirect_opts, CFGF_MULTI | CFGF_TITLE),
+		CFG_SEC ("proxy-pass", proxy_opts, CFGF_MULTI | CFGF_TITLE),
 		CFG_END ()
 	};
 	cfg_opt_t opts[] = {