+++ src/mmc.c
@@ -36,16 +36,8 @@
 #include <fcntl.h>
 #include <syslog.h>
 #include <errno.h>
-#ifdef TIME_WITH_SYS_TIME
-# include <sys/time.h>
-# include <time.h>
-#else
-# ifdef HAVE_SYS_TIME_H
-#  include <sys/time.h>
-# else
-#  include <time.h>
-# endif
-#endif
+#include <sys/time.h>
+#include <time.h>
 #include <sys/mman.h>
 
 #include "file.h"